raspit-web/requirements.txt
Flask==1.0.2
google-cloud-datastore==1.7.0
google-api-python-client
Jinja2==2.10
requests